server/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
asyncpg
strawberry-graphql[fastapi]